Added main graph file and two algorithm files
The main graph file lib/DepGraph contains the logic to retrieve and build
the dependency graph. This means a graph of nodes linked by primary
dependencies. The other two "algorithm" files in lib/algs contains mix'ins
that add varioius algorithms to the DepGraph class. The two that I have
added are Primary and Secondar Links. Right now these files are just blank,
but they are reminders of the command-pattern that I would like to use to
implement them.