Added basic ability to link dependencies in graph
I already had the logic to pull down all the gems and store them in the
graph but, I needed a way to link all of the dependencies together. This
is done now. However, I have not included logic to handle the case when
a gem list a dependency that cannot be found.