Added logging and common classes
Needed logging so that I can track the progress of the scripts as they
go out to the internet and download marshal files and build the dependency
graph. In case something errors or goes wrong I can see what's up.

It's also nice for outputing and saving the dependency graph when it is done
so that I don't have to worry about having to re-download all of the stuff
again.