Update docs/GOAL_ROADMAP.md
Updated the GOAL_ROADMAP.md file to match the updated
items in the README.md file. However, I'm questioning
the real value of this document anyways. Who am I
justifying this too anyways?? Or, perhaps it's just to
document the motivations for the project to allow other
possible committers to decide if the goals of this
project fall in line with their own goals... Who knows.
Perhaps I'll move/rename/re-purpose this document and just
put all of the roadmap stuff on Trello.