Cleaned up project
I'm picking this project back up and (hopefully) finishing within
the next few months or so. As such, I just wanted to perform some
clean-up operations such as updating the Gemfile and the .gitignore
file.

Also, one big thing, is that this should not be a gem. Honestly, I'm
not sure what I was thinking when I had originally created like
that. I'm not sure what the use case would even be for that. Anyways,
that is gone now and it should not come back.