Renaming project to Laces
The project was initially called Influential, but I don't like that
name and I feel like it's going to be more of a tool of anlysis and
influence that a gem has on the development of other gems is only
one way that the analysis engine could be used (can I call it an
engine?? lol).

Anyways, I'm going to call it laces. Mainly becuase we're looking
at how the gem universe is tied together and that just seemed fitting.
Also, there is only one gem called laces on rubygems.org and it doesn't
look to be too popular so I'm not that worried about getting people
confused. lol.