Updated Tests & Docs
The docs still had some references to 'influential' that I
have updated to 'laces'. Also, the tests weren't running because
RSpec had been upgraded and needed to make a minor change to how the
module was being loaded. All that is working now and the tests are
green.

So...I belive that I can get back to programming on this thing now
after I define a clear scope and purpose for the project.